path: use the right datatype
The strlen() function returns a size_t
Storing this in a standard signed int is a bad practice
that invites overflow vulnerabilities if paths get absurdly long.

+++ path.c
@@ -17,6 +17,7 @@
 #include "odb.h"
 #include "lockfile.h"
 #include "exec-cmd.h"
+#include <stddef.h>
 
 static int get_st_mode_bits(const char *path, int *mode)
 {
@@ -58,7 +59,7 @@ static void strbuf_cleanup_path(struct strbuf *sb)
 
 static int dir_prefix(const char *buf, const char *dir)
 {
-	int len = strlen(dir);
+	size_t len = strlen(dir);
 	return !strncmp(buf, dir, len) &&
 		(is_dir_sep(buf[len]) || buf[len] == '\0');
 }
